bd: Switch to NUMERIC instead of INTEGER

+++ TopicosEmBancosDeDados/endereco.sql
@@ -6,7 +6,7 @@ CREATE TABLE UF(
 );
 
 CREATE TABLE MUNICIPIO(
-    Codigo_Municipio INTEGER NOT NULL,
+    Codigo_Municipio NUMERIC(5) NOT NULL,
     Nome_Municipio CHARACTER VARYING(80) NOT NULL,
     Sigla_UF CHARACTER VARYING(2) NOT NULL,
     
@@ -15,38 +15,38 @@ CREATE TABLE MUNICIPIO(
 );
 
 CREATE TABLE BAIRRO(
-    Codigo_Bairro INTEGER NOT NULL,
+    Codigo_Bairro NUMERIC(5) NOT NULL,
     Nome_Bairro CHARACTER VARYING(80) NOT NULL,
-    Codigo_Municipio INTEGER NOT NULL,
+    Codigo_Municipio NUMERIC(5) NOT NULL,
 
     CONSTRAINT PrkBairro PRIMARY KEY (Codigo_Bairro),
     CONSTRAINT FrkMunicipioBairro FOREIGN KEY (Codigo_Municipio) REFERENCES MUNICIPIO (Codigo_Municipio),
 );
 
 CREATE TABLE LOGRADOURO(
-    Codigo_Logradouro INTEGER NOT NULL,
+    Codigo_Logradouro NUMERIC(5) NOT NULL,
     Nome_Logradouro CHARACTER VARYING(80) NOT NULL,
     CEP CHARACTER VARYING(8) NOT NULL,
-    Codigo_Bairro INTEGER NOT NULL,
+    Codigo_Bairro NUMERIC(5) NOT NULL,
 
     CONSTRAINT PrkLogradouro PRIMARY KEY (Codigo_Logradouro),
     CONSTRAINT FrkBairroLogradouro FOREIGN KEY (Codigo_Bairro) REFERENCES BAIRRO (Codigo_Bairro),
 );
 
 CREATE TABLE ENDERECO(
-    Id_Endereco INTEGER NOT NULL,
+    Id_Endereco NUMERIC(5) NOT NULL,
     Quadra CHARACTER VARYING(20) NOT NULL,
     Lote CHARACTER VARYING(20) NOT NULL,
     Numero CHARACTER VARYING(20) NOT NULL,
     Complemento CHARACTER VARYING(255) NOT NULL,
-    Codigo_Logradouro INTEGER NOT NULL,
+    Codigo_Logradouro NUMERIC(5) NOT NULL,
 
     CONSTRAINT PrkEndereco PRIMARY KEY (Id_Endereco),
     CONSTRAINT FrkLogradouroEndereco FOREIGN KEY (Codigo_Logradouro) REFERENCES LOGRADOURO (Codigo_Logradouro),
 );
 
 CREATE TABLE TELEFONE(
-    Id_Endereco INTEGER NOT NULL,
+    Id_Endereco NUMERIC(5) NOT NULL,
     Telefone CHARACTER VARYING(20) NOT NULL,
 
     CONSTRAINT PrkTelefone PRIMARY KEY (Telefone, Id_Endereco),
@@ -64,7 +64,7 @@ CREATE TABLE PESSOA(
 );
 
 CREATE TABLE TIPODEENDERECO(
-    Codigo_Tipo_End INTEGER NOT NULL,
+    Codigo_Tipo_End NUMERIC(2) NOT NULL,
     Nome_Tipo_End CHARACTER VARYING(20) NOT NULL,
 
     CONSTRAINT PrkTipoDeEndereco PRIMARY KEY (Codigo_Tipo_End),
@@ -72,8 +72,8 @@ CREATE TABLE TIPODEENDERECO(
 
 CREATE TABLE POSSUI(
     cpf CHARACTER VARYING(11) NOT NULL,
-    Id_Endereco INTEGER NOT NULL,
-    Codigo_Tipo_End INTEGER NOT NULL,
+    Id_Endereco NUMERIC(5) NOT NULL,
+    Codigo_Tipo_End NUMERIC(2) NOT NULL,
 
     CONSTRAINT PrkPossui PRIMARY KEY (cpf, Id_Endereco),
     CONSTRAINT FrkPessoaPossui FOREIGN KEY (cpf) REFERENCES PESSOA (cpf),
